+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - `tabs` in `tabs` view
 - Changed views to normalize props for render: `checkbox`, `expand`, `tabs` and `tab`
 - Added `Model#getContext()` method
+- Removed `ViewModel#queryToConfig()` method
 
 ## 1.0.0-beta.93 (06-02-2025)